Updating heading for reasons for rejection page
This aims to make it clearer that the rejection reasons are being sent to the candidate and should be directed at them.

We have evidence that some providers do not currently understand this, as they are writing about the candidate in the third person and using very short, terse reasons such as.