Add missing explanation header if candidate does not have GCSE
This label needs to be conditional upon whether the candidate answered "Yes" or "No" to "Are you currently studying for this qualification?"

If they answered Yes, and then completed the free text, then the label should be "Details of qualification currently studying for". If they answered No, and then fill out the "If you have other evidence of having maths skills at the required standard, give details (optional)" then we should display "Other evidence I have the skills required" with their answer.