Split up `ProcessState`
It was doing two sort-of related jobs; inferring the state of
application_forms, and providing a candidate workflow for flow diagrams.

After discussion with the team I've split this in two and provided what
are hopefully clearer names, and removed the "experimental" warning
since it's in use in the candidate API and some data exports (as well as
the original use in the support UI).