Convert the v1.2 candidate API over to using standard `updated_at`s
This is "good enough" to what the API is used for without requiring us
to maintain a custom timestamp which (as evidenced) is hard to maintain.

I'm not certain this level of complexity is necessary as I think we
could ditch some of the `ORDER BY` components and thereby maybe stop
using a subquery.

It's not super clear why the candidates need to be in reverse updated
order but the forms should be in normal created order.