Ignore duplicate site codes for archived sites
If a site is removed from Publish, we mark any CourseOptions associated
with it as `site_still_valid: false`. When it comes to validating an
application choice, we do not wish to raise an error if the selected
site code matches a site which has effectively been archived (ie removed
from Publish).