Build Satisfactory types automatically from FactoryBot definitions
Still some fine tuning to go, and could do with a neater hook for
custom traits/associations (needs moving into app code).

Also not convinced about keeping the dynamic-methods part of the DSL,
might be worth a reduction in usage readability in exchange for an
increase in library code readability.