Remove support 2022 and older cycles for reason for rejection dashboard
The report in support interface for 2022 was innacurate,
wrong and incomplete.
Now that it was reported the bug, maintain both structures for 2022
alone would be not viable because we would need tons of different
codes to maintain different structures of JSON in the DB while need
to group and compile the information for 2022 in order to be
accurate. Now that the structure is concise,
we should follow for 2023 and beyond.