Restructure Terraform to separate PaaS and AKS
The previous structure meant a condition was required in the module
selection and this changed the resource names in PaaS. This led to
Terraform wanting to destroy the existing infrastructure. Updating to
separate out AKS and PaaS and added required Makefile targets
so we can deploy either one or the other.