Add validations for ApplicationForm
A candidate is only able to sign up for an adviser when certain criteria
are met; they must have provided the required candidate information and
have relevant qualifications.

All candidates must provide:

- Email
- First/last name
- Date of birth
- Telephone number
- Nationality

UK candidates are also required to have:

- Postcode
- GCSE maths, english (passed or retaking)
- A UK degree (Bachelor's and above, at least 2:2).

International candidates are required to have:

- UK equivalent of Bachelor's degree

In the scenario where candidates may have added multiple suitable degrees
we are taking the highest-graded UK degree (international degrees are
unsorted).

We are treating international candidates as those with an applicable
international degree (rather than using their nationality) which is
consistent with the 'Get an adviser' service.

I opted to filter the degrees in-memory rather than as part of the
initial query for simplicity and on the assumption that the majority of
candidates only have one degree.