Add model to sign up for an adviser
Add a model to sign up for an adviser; in order to sign up the
`ApplicationForm` must be in an applicable state and the relevant feature
must be enabled.

As the standard `ApplicationForm` does not ask the user what their
preferred teaching subject is we need to collect this information as part
of the submission. We verify it against the list of subject items in the
GiT API/CRM.