Add job to sign up for an adviser
The job accepts a pre-validated `application_form` and
`preferred_teaching_subject_id`. It is responsible for constructing a
suitable/valid request payload from the provided details and submitting
it to the Get into Teaching API.

It is possible that the `ApplicationForm` associated with the job could
mutate into a state that is no longer applicable for an adviser if the
candidate updates the details while the job is in the queue. I would
expect this to be very edge-case as most jobs will execute fairly
instantaneously and the candidate would likely have to undo a part of
their application form. If we find that we do see errors from the sign up
submission we could expand the job to check for eligibility and notify
the candidate in this scenario.

There is no matchback logic implemented yet; all submissions will be
treated as new candidates in the CRM.