Add signed_up_for_adviser attribute to ApplicationForm
We need to be able to flag an application form when a candidate has signed
up for an adviser so that we no longer surface the adviser call to action to
them.

Add `signed_up_for_adviser` field to `ApplicationForm` and default to
`false`. The field will be kept up to date with the GiT API candidates at
the point of use so we don't need to backfill the column.