Perform adviser status check
When we check if the adviser sign up is available we want to first ensure
the `signed_up_for_adviser` attribute is up to date with the candidate
information stored in the GiT API.

To do this we perform a 'matchback' by submitting details of the Apply
candidate and checking if a matching candidate can be found in the GiT API.
If found, we update the adviser status based on the response.

We store the status check result from the GiT API in the Rails cache for 30
minutes to prevent multiple calls to the matchback endpoint as the user fills
out their application form.

If the GiT API is unavailable the adviser sign up will be hidden and a
warning will be logged.

Once a candidate has signed up for an adviser they become ineligible.