Update AdviserSignUpWorker to send matchback information
If a candidate already exists in the GiT API we need to fill out additional
attributes in the payload to sign up for an adviser. These are the
`candidate_id` of the existing candidate, the `qualification_id` (so the API
can update the most relevant qualification with the new information) and
`adviser_status_id` (used by the API to perform adviser reassignment).