Update descriptions of GCSE-equivalent qualifications
Evidence from validation error logs show that a lot of users are seeing an error message when adding an O level with a qualification year after 1988.

Most of these are due to people adding qualifications from Nigeria, which still has O levels today.

We think people are doing this as they don't read the full list of options and then realise that there is a separate section for non-UK qualifications.

To improve this, we’ve relabelled "O level" to "UK O level (from before 1989)" to make it clearer that this option is only for UK O levels.

We've also relabelled "Non-UK qualification" to "Qualification from outside the UK" to distinguish it more from "Another UK qualification", and have added an additional "or" divider.