Track adviser sign up in dfe-analytics
We want to be able to analyse the success rate of the adviser sign up. In
order to do this we will need to know when candidates have qualified/been
offered the adviser service and when they go on to sign up for an adviser.

Add custom dfe-analytics events to track engagement.