Add adviser_status attribute to ApplicationForm
Currently we can only determine if a candidate has
signed up for an adviser or not. The updated design
for this journey displays a different UI to the
candidate depending on their adviser status, so we
are going to replace the `signed_up_for_adviser`
attribute with an `adviser_status` attribute. The
possible values will be:

```
enum adviser_status: {
  unassigned: 'unassigned',
  waiting_to_be_assigned: 'waiting_to_be_assigned',
  assigned: 'assigned',
}
```

Once we have updated the availability logic to use
this new field we can remove the
`signed_up_for_adviser` attribute.