Switch sign up availability to check adviser_status
Currently when we check if an `ApplicationForm` is eligible for an adviser
we reference the `signed_up_for_adviser` attribute. We have since added a
field to store the `adviser_status` in more detail; a candidate can be in
one of these adviser states:

```
unassigned
waiting_to_be_assigned
assigned
previously_assigned
```

The `unassigned` state is equivalent to not being `signed_up_for_adviser`.
We will use the additional states in a follow up change to provide more
tailored guidance to the candidate.