Update SignUpAvailability cache on adviser_status change
There is coupling between an `application_form.adviser_status` and the
`SignUpAvailability` class. If the `adviser_status` is updated outwith this
class it may have unexpected results as the value as cached by
`SignUpAvailability` will not be updated. To avoid this I've opted to add an
`after_save` hook to the `ApplicationForm` model that will notify the
`SignUpAvailability` of the change/allow it to update its cached value.