Add TeachingSubjects class
We need to be able to retrieve teaching subjects from the API and filter
them to those that we want to show in Apply. Akwardly, there is also a
'primary' teaching subject that we want to present differently so as not to
confuse the user.

Add `TeachingSubjects` service to encapsulate this logic.