Expand SignUpAvailability to check adviser status
If the sign up is not available for the candidate we want to display a
different message to them depending on if they are waiting to be assigned or
if they have already been assigned to an adviser.

Expand `SignUpAvailability` with methods to determine the adviser assignment
status. We perform the pre-checks on all methods to ensure we have the
up-to-date information from the API.