Fix sorting bug when index of grade is nil
If a candidate has both domestic and international degrees the domestic
degree grades will result in a numerical index that can be sorted whereas
the international degree grade could evaluate to `nil` as its a free text
input. If `min_by` tries to sort a collection containing `nil` it will raise
an error; instead, we want to return an index value greater than all the
domestic degree grades so that international degrees take less precedence.