Check address type on requiring postcode
For the purposes of the adviser sign up we currently classify an
international candidate as one with an international degree. This makes
sense when checking if we require GCSEs (only candidates with a domestic
degree require GCSEs), however we need to check the address type is
international/domestic when determining if a postcode is required.

Currently, if a candidate provides an international address and domestic
degree they will not qualify for an adviser (and they should). This is
because they have not/cannot enter a postcode, which is required. By
changing the postcode validation to check the address type it correctly
matches up with the existing validation in Apply around when the postcode
field is offered.