Fix logic around GCSE status for adviser sign up
When checking to see if a candidate is eligible for an adviser we ensure
that they either have or are retaking GCSE Maths and English. When we come
to submit the sign up payload to the API, however, it becomes more
complicated as the API/CRM combines Maths and English status into single
fields.

This has caused an edge case issue whereby a candidate may, for example,
have a Maths GCSE and is retaking their English GCSE. As it stands, this
would result in `false` being passed to both the
`has_gcse_maths_and_english_id` and
`planning_to_retake_gcse_maths_and_english_id` attributes, resulting in a
400 bad request response from the API.

We can handle this scenario a number of ways and after discussing with the
CRM team we have decided that if the candidate has one GCSE and is retaking
the other we should send `false` for `has_gcse_maths_and_english_id` and
`true` for `planning_on_retaking_gcse_maths_and_english_id`.