Fix serialisation of free text grades
Previously the forms for 'other' degree grades that don't have a
specific list of allowed grades were not working correctly. When the
candidate selected _Yes_ (I do have a grade) and then entered the grade
in the revealed text input field we serialised `'Yes'` as the grade
rather than the contexts of the input field.