Attempt to fix flakey test in delete application spec
This spec asserts that first and last name are not present (anywhere) on
the page. There is a risk that the Faker generated names will match
another piece of legitimate text on the page thus failing the spec.

I've attempted to prevent this happening by wrapping the assertions in a
`within_summary_row` so that we only look for the names in relevant
summary boxes.