Dependency so webapp is deployed before workers
There is a large number of this Sentry error: ActiveRecord::NoDatabaseError
It only happens with review apps. We suspect that when the workers start, there is no database and sidekiq complains.
The database is created by the db:prepare command on the webapp and may take some time to create.

To work around this issue, we should precreate the database before the workers start.

This is achieved by creating a dependency between the webapp and the workers.