Flakey test in OrganisationPermissionsExport
Attempting to fix https://github.com/DFE-Digital/apply-for-teacher-training/actions/runs/5253338097/jobs/9490893914?pr=8318

It's hard to reproduce this failure (couldn't do it locally). The error
is a little hard to decipher due to the escape sequences in the output
but what it boils down to is that the assertion
`expect(provider_code).to be_nil` fails even though the 'current user'
in the audit trail is a support user.

Looking at the query there seems to be a loophole in the join
`LEFT JOIN provider_users_providers pup ON audits.user_id =
pup.provider_user_id`. To avoid getting joining to users that are not
provider users we need to put a condition on the `user_type` too.