Introduce invariant check for the monthly stats report
The monthly statistics report is generated w week before it is
published. On a couple of occasions the report has not been
generated and we are only alerted to it not appearing thanks to
members of the public. This invariant check will alert us if the
report has not been generated.