Fix bug where the error summary is presented on application review
Because we were centralizing everything in one object form
when we call valid? the errors were added and the error summary
was shown.

For that matter, I decided to separated into two objects

SubmitApplicationForm -> handle the yes/no answer and renders the form
ApplicationChoiceSubmission -> verify if the application choice is
ready for submission

Be mindfull that the ApplicationChoiceSubmission **is not** a form
object therefore I did move to the services dir