New personal statement journey (#8451)
## Context

Implement the new Personal Statement Journey.

With continuous applications, there is a new personal statement per
application choice. The personal statement will be copied from the
`ApplicationForm#becoming_a_teacher` when the application choice is
submitted. Candidates must update the
`ApplicationForm#becoming_a_teacher` field before each application
choice submission in order to customise the personal statement for that
application choice.

Before continuous applications, when an application form was `submitted`
the personal statement could not be reviewed or edited. Now in
continuous applications, the application form becomes submitted after
the first application choice is submitted. But the candidate needs
access to edit the personal statement in order to customise it for the
next application choice.

We allow the candidate to edit personal statement if the application
form is submitted and the application form is continuous applications.

## Changes proposed in this pull request

### Vendor API

All versions of the Vendor API will return the `personal_statement` that
exists on the `ApplicationChoice` model rather than the values of
`becoming_a_teacher` and `subject_knowledge` that exist on the
`ApplicationForm` model for applications in 2024 recruitement cycle
(continuous applications)

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->

## Link to Trello card


https://trello.com/c/9XnY0FJ3/38-ca-build-the-new-personal-statement-journey

Partially:
https://trello.com/c/tUQa5Eus/36-ca-allow-application-form-sections-to-be-editable

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] This code does not rely on migrations in the same Pull Request
- [ ] If this code includes a migration adding or changing columns, it
also backfills existing records for consistency
- [ ] If this code adds a column to the DB, decide whether it needs to
be in analytics yml file or analytics blocklist
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Required environment variables have been updated [added to the
Azure KeyVault](/docs/environment-variables.md#deploy-pipeline)