Backlinks updated for continuous applications (#8471)
## Context

Dynamic back links for continuous applications candidate interface


## Changes proposed in this pull request

1. Move the "Backlinks" methods out of `ViewHelper` and into a concern
and include them in `CandidateInterfaceController`.
2. Add `application_form_path` method which will return a path based on
whether the user is in continuous applications or not. If continuous
apps then point to Your Details page, otherwise point to current
application form url.
3. Update `govuk_back_link_to` method to automatically apply "Back to
your details" or "Back to your applications" based on the the url
pattern.

There is an issue with the Backlinks `application_form_path` method.
The method calls `current_application`. This is a method defined on
`CandidateInterfaceController` and made into a `helper_method` there.
This method does not become a `helper_method` in isolated view specs
however. I've defined the method in `ViewHelper` to fix this. I don't
know how else do manage this for now.


Mostly the PR swaps out `candidate_interface_application_form_path` for
`application_form_path`.