Pluralise Your applications title (#8475)
The Primary Navigation contains items Your details and Your
application(s). If the candidate has none or one application at least in
draft, the navigation item displays singularly. If there is more than
one, it pluralises "application".

We've decided it's better to just keep it pluralised even if there's
only one or none.