Add content message about your details are incomplete
When a candidate has not completed all sections of ‘Your details’, and they go to the ‘Your applications’ tab, we show the following content:

You can add up to 4 applications at a time.

You will not be able to submit applications until you have completed your details.

The second line disappears if all sections are completed in ‘Your details’