Filter out repetitive site_setting-related log messages
This commit adds a new initializer that filters out repetitive log
messages. This is to reduce the noise in the development log and make it
easier to spot other log messages.

The initializer is written in such a way that it accepts a list of
strings to filter out. This means that we can add more
filters in the future if we need to.

The `RAILS_FILTER_STDOUT` environment variable is used to enable this
feature. It must be set to `true` for the filter to be applied.