Highlight the primary navigation items based on controller path (#8477)
## Context

The primary navigation items are split between "details" and
"applications"
  for continuous applications. All general application form details
conceptually belong to the "details" tab. All application choice
specific
  pages belong to the "applications" tab.

  Create a method that decides which tab should be displayed as active
  based on the current requests controller path.

  For legacy application form, only one tab exists, so the decision
  method returns false always. if continuous applications, check the
  current controller path against a list of matches.

  For guidance, this page can be reached from either tab. if the
  controller is guidance we check the referer matches choices.

## Changes proposed in this pull request

Add `choices_controller?` method to
`CandidateInterface::CandidateInterfaceController`. This is used to
determine whether the controller processed by the current request should
be considered as happening within the "Your applciations" tab in the
Candidate Interface.

This method is used to make the Primary Navigation item appear active.
If it's not active, then the "Your details" tab is shown as active.