Show Previous application heading if submitted (#8500)
## Context

Previous applications heading appears in the right sidebar on the Your
applications page for legacy applications. In Continuous applications,
this component will not be required. Previous application refers to
applications from which the current application has been carried
over/apply again. There is no such concept in continuous applications.

## Changes proposed in this pull request

Only show the heading if there are previous applications and any of them
are submitted. Currently, the heading shows if there are any previous
applications, but only applications that are submitted are listed under
the heading.

Now we only show the heading if there are any applications to be listed
under the heading.