Set correct next_edit_step_path for DuplicateCourseSelectionStep
next_edit_step_path should be next_step_path for duplicate_course_selection.
This value is used to determine the next step when editing the course
selection. When there is a duplicate validation error during edit, we
want the next step path to be the 'new' action rather than the 'edit'
action for DuplicateCourseSelectionStep