Disappear the "have you complete this section" on editable sections
Last minute request was to disable this section on editable sections

The reason for the condition:
if submitted_applications? && form.object.completed.present?

Is because of science GCSE and english as foreign language that still
needs to have the question even after any application is submitted.

Created this card as consequence of this commit:
https://trello.com/c/6iWzA6QE/710-science-gcse-english-as-foreign-language-have-you-complete-this-section-question

Also I took the decision to use view components 'helpers' because
otherwise I would have to change the parameters of the component in all
sections