Fix the bug 'Candidate is unable to navigate back to the post offer'
before this commit Clicking ‘Back to application’ takes the candidate
back to the old pre-continuous applications ‘Your applications’ view

A candidate should be able to view their submitted application – this is correct.

However, it should be clear, and easy, for them to get back to the post
offer dashboard.

Ensure the back link takes them back to the post offer dashboard and
update it to say ‘Back to your offer’

We also need to ensure the ‘Date submitted’ value is using the
sent_to_provider_at attribute on the application choice,
and not the submitted_at application form attribute

We should also ensure the back link on the withdrawal page
correctly redirects to the post offer dashboard: