Only show the inset if the applications is unsubmitted and course full
When a candidate has added a course that becomes full (and the application
is unsubmitted), we used to render some text to say that it’s full.

This made use of a blue inset text.

We appear to have ported this over to our designs, but we also appear
to be rendering some sort of blue inset, regardless of the status
this means that the blue inset bar is rendering for submitted apps,
without any helptext.

This commit fix to remove this blue inset if the application is
submitted.