Add option to modify individual jobs on application forms
This commit adds a "Change" link to the work history section of an
application form. This link takes the support user to a new page where
they can edit the details of a single job. It reuses the candidate
interface form for editing jobs, but with a different controller and
view.