Add option to modify individual volunteering roles on application forms
This commit adds a "Change" link to the work history section of an
application form. This link takes the support user to a new page where
they can edit the details of a single volunteering roles. It reuses the
candidate interface form for editing volunteering roles, but with a
different controller and view.