Introducing support functionality to grant an editable extension
A support user will be able to click a button to ‘make the application editable’.
On a new page, they will then be able to set a non editable section
temporarily editable and the code will automatically calculate the
date until the candidate will be able to edit.
This will set the editable_until value.
We’ll also request the Zendesk link of the originating request.
Once set, the support user will return to the candidate’s application
view and will be able to see the ‘editable until’ value rendered.
when editable_until > current date, the candidate can edit all sections
selected via support.