Implement a simple SQL where order
This is to help maintain the queries from Bigquery is a sane matter.

I tried to use Arel::Table but the problem is that arel table
returns the where clauses with the table name
"dataform"."application_metrics"."colum_name" and this raises
and invalid query for bigquery.

Tried overwriting the "from" in Arel::Table also doesn't help

So only option was to write a simple SQL kinda of DSL