Separate logic between group date for report and group for CSV
Before I was using the grouped attributes for report to generate the CSV
but that was proving to be very complicated.

So I decided to loop through big query records to use as rows
which is just a simple loop.

And made a bigquery title column so we can map each section into the
report and the extra columns for each of those sections