Remove caching on provider selection pages
This is causing issues because it only checks the provider updated
values, but not the course updated values. So when the courses get
published, the cache doesn't break and the provider doesn't show up.
Instead of trying to fix this, I'm just removing the cache.
Doesn't seem to be a performance issue.