Show correct country for international qualifications
Fixes the bug where international qualifications were showing as UK.
I am only changing the way it gets displayed in the support interface,
although the ticket suggests that the `international` boolean should be
true. It does seem to be used for degrees, but not for other
qualifications. If we decide to use it for other qualifications, we
should add a migration to set it to true for all existing international
qualifications.