[1021] Populate `work_history_status` when nil
Some carried over applications have a `work_history_status` of nil. This is because they have been created before the `work_history_status` was added. This is preventing the "Add another job" button from rendering.

We fixed the issue in: https://github.com/DFE-Digital/apply-for-teacher-training/pull/8846

But we need to backfill the values for the applications forms that already have it set to nil.