Remove the upload of artifacts on GitHub workflow builds
The artifacts upload is deprecated and has to be updated, but because it's not useful for us, we can just remove it.

I added it so we can debug flakey specs, unfortunately, it only works if the tests is running with `:js`. We have very few specs running with `:js` and it's not worth the effort to keep it.