[1056] Populate institution_country when nil and uk degree
There was a bug in the code which caused institution_country to be set to nil when adding a UK degree. It should have been set to 'GB.

A fix for the logic was done in this PR: https://github.com/DFE-Digital/apply-for-teacher-training/pull/8956

This is a data migration to fix the records which are already in production.