Fix submission bugs
It is possible to by pass submission if you add primary and secondary
course where the science gcse is not needed validation was checking
with an "&" statement.

To avoid this error prone scenarios I decided to move the
incomplete validators down the line and we should validate ALL the
sections are complete as last resort to avoid any possible scenario
to bypass the submission.