Make sure carry over does work in all scenarios
Before this commit all applications were pointing to the 2023
cycle controller.

Now that this controller is gone there is a need to move some filters.

So I created two filters:

unsuccessful_carry_over_filter.rb - for the submitted and unsuccessful
applications we need to carry over
carry_over_filter.rb - the unsubmitted apps that reached the end of the
cycle

Also the complete route should render for 2023 and 2024 (2024 only
    if the application is not carried over)