Try to fix flakey spec
Getting `PG::ForeignKeyViolation: ERROR:  insert or update on table
"application_forms" violates foreign key constraint "fk_rails"` error
when running on CI. It seems the a new `@application_form` gets created
on every scenario in that spec, but it uses the same `current_candidate`
variable. This might be causing the issue.