Return unauthorised when authentication token does not exist
We have Sentry errors, indicating issues with Register <> Apply API
integration. These are being posted to our _tech channel.

undefined method `user_id' for nil:NilClass (NoMethodError)

This is because when a wrong authentication token is passed, we
were not verifying if exists or not. This commit solves this issue.