Add a document to outline the various Apply APIs
I've written up a high level document covering the four different APIs
that Apply provides, including the intended use and audience of each.

I found it quite hard to uncover and locate documentation for the
various APIs this app has, so I'm hoping this document will serve as a
helpful overview.

This follows a lengthy Slack thread of discovery:
https://ukgovernmentdfe.slack.com/archives/C04MLBVP876/p1711020827068249?thread_ts=1710927942.265709&cid=C04MLBVP876