Change semantic logger to use in development and JSON in production
As one team in Find/Publish/Apply/Manage we should have the same
way of handling logs in development so adding the semantic logger
on Apply.

I tried to add this to the semantic logger but it could not remove
the semantic logger by an ENV var.

+if ActiveModel::Type::Boolean.new.cast(ENV['RAILS_DEFAULT_LOGGER']).present? && HostingEnvironment.development?
+  Rails.application.config.semantic_logger = nil # Disable semantic logger completely
+  Rails.application.config.rails_semantic_logger = OpenStruct.new
+
+  Rails.logger = Logger.new(STDOUT)
+    .tap  { |logger| logger.formatter = ::Logger::Formatter.new }
+    .then { |logger| ActiveSupport::TaggedLogging.new(logger) }
+  return
+end
+

The reason is that the engine from rails semantic logger added
many "hacks" to rails replacing the logger and monkey patching
active support logger. So there is no way as right now to disable
semantic logger in developmet using an env var