Implement zendesk messaging
Zendesk Chat service is being deprecated in favour of their Messaging
service.

This PR implements the new Messaging service, in a very similar way to
the get into teaching website implemented it

With a link in the footer which will open the messaging widget.

We cannot get the opening hours of the zendesk messaging in our app, like
the old implementation, meaning we cannot change the messaging on our
website based on the schedule.

The user should be informed of the working hours of the support team
after opening the messaging widget.

A zendesk messaging key needs to be added to the env.test file so that
the tests pass. This key is a sandbox key so it should be fine.