Reference wizard, views and base controllers
The routes have been consolidated under /references/:reference_process
where reference process refers to the different type of reference.
Consolidating the controllers as well, removing the need for inheritance
for the accept-offer and request-reference references.

candidate-details the reference that the user adds as a candidate
accept-offer the reference the user adds after accepting the offer
request-reference the reference the user requests

The dfe wizard gem has been used to handle the multi step form for every
wizard, with 4 steps(type, name, email and relationship). The dfe wizard
steps also handle the redirection to the next or previous steps. I didn't
use the next_step_path helper that the dfe wizard provides as I need to
specify the paths manually and they need to be more dynamic than the gem
allows.