Added ReferencePresenter from ApplicationPresenter
The Reference data was previously rendered as part of the Application Presenter.
This has been moved out of the ApplicationPresenter into its own Presenter.

This Presenter is added to the v1.0 version as this is where it was first introduced.

In v1.3 there were additions to the Reference Presenter which have been refactored as an API change.