Reduced the number of records being returned in email_log_controller.rb.
Added a new `days_ago` filter that reduces the number of Email we attempt to return.

Removed second Email query from the view to display the first created_at date. This have been replaced by the `created_since` in the `applied_filter` which is built based on the `days_ago` parameter.