Ignore ApplicationQualification#equivalency_details
This column is not used in production it's nil for most records. There
are only five records with any data in this field – and the data is Lorem
Epsom, so test data rather than actual qualification data.

This column is not used in any form. It will be removed in a future PR.