Added the reference to the formatted description on English Proficiency (#9666)
## Context

The `EnglishProficiency` model has a method to render out a formatted
description of the attached qualification.
These qualifications can be an `IeltsQualification`, a
`ToeflQualification` or an `OtheEflQualification`.
All but the `OtherEflQualification` have a unique reference. 

A Vendor API consumer has requested that we add the reference to the
formatted description so that they can populate data in their system.

A full solution to this issue would be to add additional attributes to
the API object. We may do this in future.

## Changes proposed in this pull request

- Added reference data to
`EnglishProficiency#formatted_qualification_description`