Ignore ApplicationExperience#application_form_id
The application_experience table should have all the experienceable
columns populated meaning we can rely on those columns to link it to the
application_form.

This commit ignores the application_form_id, another commit will remove
this column.

We also need to remove the data migrations that use this column.

And we need to explicitly delete the application work experiences in
`delete_test_applications.rb` because having polymorphic associations
will not delete the work experiences when deleting the candidate, it
doesn't do cascade deleting

Same issue in the candidate_spec file