Make sure mobile users don't experience Authenticty token issues
We have an issue that arises because mobile Safari reloads pages
from cache after the browser is closed and reopened.
When this happens, the session cookie that stores the CSRF token
is gone, but the form still contains the old token.

This mismatch leads to an error when the user tries to submit the form.
Rails stores the authenticity token in a session cookie, which gets
deleted when the browser is closed.

On mobile Safari, when you reopen the browser, it reloads tabs from
cache, but the session cookie is gone.

As a result, the authenticity token in the form can't be verified
when the form is submitted.