Add ApplicationWorkHistoryBreak#breakable
We need to allow the ApplicationWorkHistoryBreak records to be edited by
the user after submitting the application.

We will use the same approach that we use on ApplicationExperiences, a
polymorphic column that will link the ApplicationWorkHistoryBreak to an
application_form and an application_choice.

This commit just adds the polymorphic columns, commits will follow to
use these columns